fix APath::move no such file or directory

+++ aui.core/src/AUI/IO/APath.cpp
@@ -439,7 +439,14 @@ time_t APath::fileModifyTime() const {
 }
 
 void APath::move(const APath& source, const APath& destination) {
-    destination.removeFile();
+    if (source.isRegularFileExists()) {
+        if (destination.isRegularFileExists()) {
+            try {
+                destination.removeFile();
+            } catch (...) {}
+        }
+    }
+
 #if AUI_PLATFORM_WIN
     if (MoveFile(aui::win32::toWchar(source.c_str()), aui::win32::toWchar(destination.c_str())) == 0) {
 #else
